+++ tests/TestCase/ArrayFunctionsTest.php
@@ -23,60 +23,6 @@ use PHPUnit\Framework\TestCase;
 class ArrayFunctionsTest extends TestCase
 {
     /**
-     * @group legacy
-     * @test
-     * @uses \array_clean()
-     */
-    public function testArrayClean(): void
-    {
-        $filterMethod = fn($value): bool => $value && $value != 'third';
-
-        $array = ['first', 'second', false, 0, 'second', 'third', null, '', 'fourth'];
-        $this->assertSame(['first', 'second', 'third', 'fourth'], array_clean($array));
-        $this->assertSame(['first', 'second', 'fourth'], array_clean($array, $filterMethod));
-
-        //With a string as `$callback`
-        $array = ['string', 1];
-        $this->assertSame(['string'], array_clean($array, 'is_string'));
-
-        $array = ['a' => 'first', 0 => 'second', false, 'c' => 'third', 'd' => 'second'];
-        $this->assertSame(['a' => 'first', 0 => 'second', 'c' => 'third'], array_clean($array));
-        $this->assertSame(['a' => 'first', 0 => 'second'], array_clean($array, $filterMethod));
-
-        $expected = ['a' => 'first', 1 => false, 'c' => 'third', 'd' => 'second'];
-        $this->assertSame($expected, array_clean($array, $filterMethod, ARRAY_FILTER_USE_KEY));
-    }
-
-    /**
-     * @group legacy
-     * @test
-     * @uses \array_has_only_numeric_keys()
-     */
-    public function testArrayHasOnlyNumericKeys(): void
-    {
-        $this->assertTrue(array_has_only_numeric_keys(['first', 'second']));
-        $this->assertTrue(array_has_only_numeric_keys([]));
-        $this->assertFalse(array_has_only_numeric_keys(['a' => 'first', 'b' => 'second']));
-        $this->assertFalse(array_has_only_numeric_keys(['first', 'b' => 'second']));
-    }
-
-    /**
-     * @group legacy
-     * @test
-     * @uses \array_unique_recursive()
-     */
-    public function testArrayUniqueRecursive(): void
-    {
-        $array = [
-            ['first', 'second'],
-            ['first', 'second'],
-            ['other'],
-        ];
-
-        $this->assertSame([['first', 'second'], ['other']], array_unique_recursive($array));
-    }
-
-    /**
      * @test
      * @uses \array_value_first()
      */