+++ tests/TestCase/GlobalFunctionsTest.php
@@ -18,7 +18,6 @@ namespace Tools\Test;
 use App\ExampleChildClass;
 use App\ExampleClass;
 use PHPUnit\Framework\TestCase;
-use stdClass;
 use Tools\TestSuite\TestTrait;
 
 /**
@@ -27,29 +26,6 @@ use Tools\TestSuite\TestTrait;
 class GlobalFunctionsTest extends TestCase
 {
     use TestTrait;
-
-    /**
-     * @group legacy
-     * @test
-     * @uses array_to_string()
-     */
-    public function testArrayToString(): void
-    {
-        $this->assertSame('[\'a\', \'1\', \'0.5\', \'c\']', array_to_string(['a', 1, 0.5, 'c']));
-        $this->assertSame('[]', array_to_string([]));
-
-        $ToStringClass = new class {
-            public function __toString()
-            {
-                return __CLASS__;
-            }
-        };
-        //This class implements the `__toString()` method
-        $this->assertSame('[\'a\', \'' . (string)$ToStringClass . '\']', array_to_string(['a', $ToStringClass]));
-
-        $this->expectExceptionMessage('Cannot convert array to string, some values are not stringable');
-        array_to_string(['a', ['b', 'c']]);
-    }
 
     /**
      * @test
@@ -78,17 +54,6 @@ class GlobalFunctionsTest extends TestCase
     }
 
     /**
-     * @group legacy
-     * @test
-     * @uses is_html()
-     */
-    public function testIsHtml(): void
-    {
-        $this->assertTrue(is_html('<b>string</b>'));
-        $this->assertFalse(is_html('string'));
-    }
-
-    /**
      * @test
      * @uses is_positive()
      */
@@ -103,51 +68,11 @@ class GlobalFunctionsTest extends TestCase
     }
 
     /**
-     * @group legacy
-     * @test
-     * @uses is_stringable()
-     */
-    public function testIsStringable(): void
-    {
-        foreach (['1', 1, 1.1, -1, 0, true, false] as $value) {
-            $this->assertTrue(is_stringable($value));
-        }
-
-        foreach ([null, new stdClass()] as $value) {
-            $this->assertFalse(is_stringable($value));
-        }
-
-        $ToStringClass = new class {
-            public function __toString()
-            {
-                return __CLASS__;
-            }
-        };
-        $this->assertTrue(is_stringable([]));
-        $this->assertTrue(is_stringable(['a', 1, 0.5, 'c']));
-        $this->assertFalse(is_stringable(['a', true]));
-        $this->assertFalse(is_stringable(['a', ['b', ['c']]]));
-        $this->assertTrue(is_stringable($ToStringClass));
-    }
-
-    /**
      * @test
      * @uses rtr()
      */
     public function testRtr(): void
     {
         $this->assertSame('my' . DS . 'folder', rtr(ROOT . 'my' . DS . 'folder'));
-    }
-
-    /**
-     * @group legacy
-     * @test
-     * @uses uncamelcase()
-     */
-    public function testUncamelcase(): void
-    {
-        foreach (['ThisIsASlug', 'thisIsASlug'] as $string) {
-            $this->assertSame('this_is_a_slug', uncamelcase($string));
-        }
     }
 }